# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/negotiations": {
      "post": {
        "operationId": "NegotiationsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser | agent | seller | buyer",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNegotiationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/NegotiationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser, agent, seller, buyer"
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "email must be an email",
                        "price must be a positive number"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Create a new negotiation",
        "tags": [
          "Negotiations"
        ]
      },
      "get": {
        "operationId": "NegotiationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/NegotiationResponseDto"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List all negotiations",
        "tags": [
          "Negotiations"
        ]
      }
    },
    "/api/v1/negotiations/{id}": {
      "get": {
        "operationId": "NegotiationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Negotiation ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/NegotiationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Negotiation not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get negotiation by ID",
        "tags": [
          "Negotiations"
        ]
      },
      "patch": {
        "operationId": "NegotiationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Negotiation ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser | agent | seller | buyer",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNegotiationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request successful"
                    },
                    "data": {
                      "$ref": "#/components/schemas/NegotiationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser, agent, seller, buyer"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Negotiation not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "email must be an email",
                        "price must be a positive number"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Update a negotiation",
        "tags": [
          "Negotiations"
        ]
      },
      "delete": {
        "operationId": "NegotiationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Negotiation ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Required role. Allowed: admin | superuser | agent | seller | buyer",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Missing or invalid role header"
          },
          "403": {
            "description": "Forbidden. Required role(s): admin, superuser, agent, seller, buyer"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Negotiation not found"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2025-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/v1/resource/123"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Delete a negotiation",
        "tags": [
          "Negotiations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Real Estate API",
    "description": "## Real Estate Backend API\n\nThis API powers the real estate platform, supporting property listings, user management, and bookings.\n\n### Authentication & RBAC\nAll protected endpoints require a `role` header indicating the caller's role.\nValid roles: `superuser`, `admin`, `agent`, `seller`, `buyer`\n\n### Response Format\nAll responses follow the envelope: `{ success, statusCode, message, data }`\n\n### Error Format\nAll errors follow: `{ success: false, statusCode, message, errors? }`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC (superuser | admin | agent | seller | buyer)"
      }
    },
    "schemas": {
...
      "NegotiationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "propertyId": {
            "type": "string",
            "example": "prop_001"
          },
          "buyerId": {
            "type": "string",
            "example": "usr_002"
          },
          "offeredPrice": {
            "type": "number",
            "example": 500000
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "propertyId",
          "buyerId",
          "offeredPrice",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateNegotiationDto": {
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "string",
            "example": "prop_001"
          },
          "buyerId": {
            "type": "string",
            "example": "usr_002"
          },
          "offeredPrice": {
            "type": "number",
            "example": 500000
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          }
        },
        "required": [
          "propertyId",
          "buyerId",
          "offeredPrice"
        ]
      },
      "UpdateNegotiationDto": {
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "string",
            "example": "prop_001"
          },
          "buyerId": {
            "type": "string",
            "example": "usr_002"
          },
          "offeredPrice": {
            "type": "number",
            "example": 500000
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          }
        }
      },
...
    }
  }
}